Show error if secret invalid
